+++ lib/external/slonik.js
@@ -7,34 +7,17 @@
 // including this file, may be copied, modified, propagated, or distributed
 // except according to the terms contained in the LICENSE file.
 
-const { readFileSync } = require('node:fs');
-
 const pg = require('pg');
-const { clone } = require('ramda');
 const { createPool, createDateTypeParser, createBigintTypeParser, createIntervalTypeParser, createNumericTypeParser } = require('slonik');
+const { withPgSslConfig } = require('../util/db');
 const { setlibpqEnv } = require('../util/load-db-env');
 
 const timestampTypeParser = { name: 'timestamp', parse: (x) => new Date(x) };
 const timestamptzTypeParser = { name: 'timestamptz', parse: (x) => new Date(x) };
 
 const { Pool } = pg;
-pg.Pool = function SslAwarePool(_cfg) {
-  const cfg = clone(_cfg);
-
-  const { PGSSLMODE, PGSSLROOTCERT } = process.env;
-
-  if (PGSSLROOTCERT) {
-    // See: https://www.postgresql.org/docs/current/libpq-ssl.html#LIBQ-SSL-CERTIFICATES
-    if (!['verify-ca', 'verify-full'].includes(PGSSLMODE))
-      throw new Error('Incompatible values for env vars: PGSSLMODE, PGSSLROOTCERT.');
-
-    cfg.ssl = {
-      ca: readFileSync(PGSSLROOTCERT),
-      rejectUnauthorized: true,
-    };
-  }
-
-  return new Pool(cfg);
+pg.Pool = function SslAwarePool(cfg) {
+  return new Pool(withPgSslConfig(cfg));
 };
 
 const slonikPool = (config) => {
