+++ lib/model/knex-migrator.js
@@ -10,28 +10,15 @@
 // This is a variety of functions helpful for connecting to and performing
 // top-level operations with a database, like migrations.
 
-const { readFileSync } = require('node:fs');
-
 const knex = require('knex');
+const { withPgSslConfig } = require('../util/db');
 const { setlibpqEnv } = require('../util/load-db-env');
 
 // Connects to the postgres database specified in configuration and returns it.
 const knexConnect = (config) => {
   setlibpqEnv(config);
-  const connection = {};
 
-  const { PGSSLMODE, PGSSLROOTCERT } = process.env;
-
-  if (PGSSLROOTCERT) {
-    // See: https://www.postgresql.org/docs/current/libpq-ssl.html#LIBQ-SSL-CERTIFICATES
-    if (!['verify-ca', 'verify-full'].includes(PGSSLMODE))
-      throw new Error('Incompatible values for env vars: PGSSLMODE, PGSSLROOTCERT.');
-
-    connection.ssl = {
-      ca: readFileSync(PGSSLROOTCERT),
-      rejectUnauthorized: true,
-    };
-  }
+  const connection = withPgSslConfig({});
 
   return knex({ client: 'pg', connection });
 };
