+++ lib/util/db.js
@@ -7,8 +7,9 @@
 // including this file, may be copied, modified, propagated, or distributed
 // except according to the terms contained in the LICENSE file.
 
+const { readFileSync } = require('node:fs');
 const { inspect } = require('node:util');
-const { mergeRight, pick, always, without, remove, indexOf } = require('ramda');
+const { clone, mergeRight, pick, always, without, remove, indexOf } = require('ramda');
 const { sql } = require('slonik');
 const { reject } = require('./promise');
 const Problem = require('./problem');
@@ -580,10 +581,30 @@ const specificFields = (F, ...fields) =>
     }
   };
 
+const withPgSslConfig = (_cfg) => {
+  const cfg = clone(_cfg);
+
+  const { PGSSLMODE, PGSSLROOTCERT } = process.env;
+
+  if (PGSSLROOTCERT) {
+    // See: https://www.postgresql.org/docs/current/libpq-ssl.html#LIBQ-SSL-CERTIFICATES
+    if (!['verify-ca', 'verify-full'].includes(PGSSLMODE))
+      throw new Error('Incompatible values for env vars: PGSSLMODE, PGSSLROOTCERT.');
+
+    cfg.ssl = {
+      ca: readFileSync(PGSSLROOTCERT),
+      rejectUnauthorized: true,
+    };
+  }
+
+  return cfg;
+};
+
 module.exports = {
   unjoiner, extender, sqlEquals, page, queryFuncs,
   insert, insertMany, updater, markDeleted, markUndeleted, sqlInArray,
   QueryOptions,
   postgresErrorToProblem, specificFields,
+  withPgSslConfig,
 };
